Move fixtures to src
