+++ src/fixtures/artists.ts
